requirements.txt
pandas 
mlflow
notebook
numpy
scikit-learn
matplotlib
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
Flask
Flask-Cors
dvc
python-dotenv
boto3
-e .